add poll_evented module for external trait impl; non-breaking refactoring of transport module
